Enum icns::IconType

pub enum IconType {
    RGB24_16x16,
    Mask8_16x16,
    RGB24_32x32,
    Mask8_32x32,
    RGB24_48x48,
    Mask8_48x48,
    RGB24_128x128,
    Mask8_128x128,
    RGBA32_16x16,
    RGBA32_16x16_2x,
    RGBA32_32x32,
    RGBA32_32x32_2x,
    RGBA32_64x64,
    RGBA32_128x128,
    RGBA32_128x128_2x,
    RGBA32_256x256,
    RGBA32_256x256_2x,
    RGBA32_512x512,
    RGBA32_512x512_2x,
}

Types of icon elements that can be decoded as images or masks.

This type enumerates the kinds of IconElement that can be decoded by this library; each IconType corresponds to a particular OSType. The non-mask IconType values can also be used with the higher-level IconFamily methods to encode and decode complete icons that consist of multiple IconElements.

Variants

RGB24_16x16

16x16 24-bit icon (without alpha).

Mask8_16x16

16x16 8-bit alpha mask.

RGB24_32x32

32x32 24-bit icon (without alpha).

Mask8_32x32

32x32 8-bit alpha mask.

RGB24_48x48

48x48 24-bit icon (without alpha).

Mask8_48x48

48x48 8-bit alpha mask.

RGB24_128x128

128x128 24-bit icon (without alpha).

Mask8_128x128

128x128 8-bit alpha mask.

RGBA32_16x16

16x16 32-bit icon.

RGBA32_16x16_2x

16x16 32-bit icon at 2x "retina" density (so, 32 by 32 pixels).

RGBA32_32x32

32x32 32-bit icon.

RGBA32_32x32_2x

32x32 32-bit icon at 2x "retina" density (so, 64 by 64 pixels).

RGBA32_64x64

64x64 32-bit icon. (For whatever reason, the ICNS format has no corresponding type for a 64x64 icon at 2x "retina" density.)

RGBA32_128x128

128x128 32-bit icon.

RGBA32_128x128_2x

128x128 32-bit icon at 2x "retina" density (so, 256 by 256 pixels).

RGBA32_256x256

256x256 32-bit icon.

RGBA32_256x256_2x

256x256 32-bit icon at 2x "retina" density (so, 512 by 512 pixels).

RGBA32_512x512

512x512 32-bit icon.

RGBA32_512x512_2x

512x512 32-bit icon at 2x "retina" density (so, 1024 by 1024 pixels).

Methods

impl IconType

fn from_ostype(ostype: OSType) -> Option<IconType>

Get the icon type associated with the given OSType, if any.

fn ostype(self) -> OSType

Get the OSType that represents this icon type.

fn is_mask(self) -> bool

Returns true if this is icon type is a mask for some other icon type.

Examples

use icns::IconType;
assert!(!IconType::RGB24_16x16.is_mask());
assert!(IconType::Mask8_16x16.is_mask());
assert!(!IconType::RGBA32_16x16.is_mask());

fn mask_type(self) -> Option<IconType>

If this icon type has an associated mask type, returns that mask type; if this is a mask icon type, or a non-mask icon type that has no associated mask type, returns None.

Examples

use icns::IconType;
assert_eq!(IconType::RGB24_16x16.mask_type(),
           Some(IconType::Mask8_16x16));
assert_eq!(IconType::Mask8_16x16.mask_type(), None);
assert_eq!(IconType::RGBA32_16x16.mask_type(), None);

fn pixel_width(self) -> u32

Returns the pixel data width of this icon type. Normally this is the same as the screen width, but for 2x "retina" density icons, this will be twice that value.

Examples

use icns::IconType;
assert_eq!(IconType::Mask8_128x128.pixel_width(), 128);
assert_eq!(IconType::RGBA32_256x256.pixel_width(), 256);
assert_eq!(IconType::RGBA32_256x256_2x.pixel_width(), 512);

fn pixel_height(self) -> u32

Returns the pixel data height of this icon type. Normally this is the same as the screen height, but for 2x "retina" density icons, this will be twice that value.

Examples

use icns::IconType;
assert_eq!(IconType::Mask8_128x128.pixel_height(), 128);
assert_eq!(IconType::RGBA32_256x256.pixel_height(), 256);
assert_eq!(IconType::RGBA32_256x256_2x.pixel_height(), 512);

fn pixel_density(self) -> u32

Returns the pixel density for this icon type -- that is, 2 for 2x "retina" density icons, or 1 for other icon types.

Examples

use icns::IconType;
assert_eq!(IconType::Mask8_128x128.pixel_density(), 1);
assert_eq!(IconType::RGBA32_256x256.pixel_density(), 1);
assert_eq!(IconType::RGBA32_256x256_2x.pixel_density(), 2);

fn screen_width(self) -> u32

Returns the screen width of this icon type. Normally this is the same as the pixel width, but for 2x "retina" density icons, this will be half that value.

Examples

use icns::IconType;
assert_eq!(IconType::Mask8_128x128.screen_width(), 128);
assert_eq!(IconType::RGBA32_256x256.screen_width(), 256);
assert_eq!(IconType::RGBA32_256x256_2x.screen_width(), 256);

fn screen_height(self) -> u32

Returns the screen height of this icon type. Normally this is the same as the pixel height, but for 2x "retina" density icons, this will be half that value.

Examples

use icns::IconType;
assert_eq!(IconType::Mask8_128x128.screen_height(), 128);
assert_eq!(IconType::RGBA32_256x256.screen_height(), 256);
assert_eq!(IconType::RGBA32_256x256_2x.screen_height(), 256);

fn encoding(self) -> Encoding

Returns the encoding used within an ICNS file for this icon type.